+++ GlacierApplication/GlacierApplication/MainPage/Model/PLI_PDTRATE.h
@@ -10,14 +10,14 @@
 
 @interface PLI_PDTRATE : SQLitePersistentObject
 @property (nonatomic,copy) NSString * PR_PDTCODE;
-@property (nonatomic,assign) float PR_PDTYEAR;
-@property (nonatomic,assign) float PR_AGE;
-@property (nonatomic,assign) float PR_SALES;
-@property (nonatomic,assign) float PR_MRATE;
-@property (nonatomic,assign) float PR_FRATE;
-@property (nonatomic,assign) float PR_MODENO;
+@property (nonatomic,assign) double PR_PDTYEAR;
+@property (nonatomic,assign) double PR_AGE;
+@property (nonatomic,assign) double PR_SALES;
+@property (nonatomic,assign) double PR_MRATE;
+@property (nonatomic,assign) double PR_FRATE;
+@property (nonatomic,assign) double PR_MODENO;
 @property (nonatomic,copy) NSString * MODIDATE;
 @property (nonatomic,copy) NSString * MODISTATE;
-+ (bool) checkNeedPR_AGE:(NSString *) aPR_PDTCODE pdtYear:(float) aPR_PDTYEAR;
-+ (bool) checkNeedPR_SALES:(NSString *) aPR_PDTCODE pdtYear:(float) aPR_PDTYEAR;
++ (bool) checkNeedPR_AGE:(NSString *) aPR_PDTCODE pdtYear:(double) aPR_PDTYEAR;
++ (bool) checkNeedPR_SALES:(NSString *) aPR_PDTCODE pdtYear:(double) aPR_PDTYEAR;
 @end
