+++ GlacierApplication/GlacierApplication/MainPage/Model/PLI_PDTRATEDIFF.h
@@ -10,13 +10,13 @@
 
 @interface PLI_PDTRATEDIFF : SQLitePersistentObject
 @property (nonatomic,copy) NSString * PR_PDTCODE;
-@property (nonatomic,assign) float PR_PDTYEAR;
+@property (nonatomic,assign) double PR_PDTYEAR;
 @property (nonatomic,copy) NSString * PR_SEX;
-@property (nonatomic,assign) float PR_AGE;
-@property (nonatomic,assign) float PR_SALES;
-@property (nonatomic,assign) float PR_RATE6;
-@property (nonatomic,assign) float PR_RATE3;
-@property (nonatomic,assign) float PR_RATE1;
+@property (nonatomic,assign) double PR_AGE;
+@property (nonatomic,assign) double PR_SALES;
+@property (nonatomic,assign) double PR_RATE6;
+@property (nonatomic,assign) double PR_RATE3;
+@property (nonatomic,assign) double PR_RATE1;
 @property (nonatomic,copy) NSString * MODIDATE;
 @property (nonatomic,copy) NSString * MODISTATE;
 @end
