+++ GlacierApplication/GlacierApplication/MainPage/Model/PLI_PDTYEAR.h
@@ -10,22 +10,22 @@
 
 @interface PLI_PDTYEAR : SQLitePersistentObject
 @property (nonatomic,copy) NSString * PD_PDTCODE;
-@property (nonatomic,assign) float PY_PDTYEAR;
+@property (nonatomic,assign) double PY_PDTYEAR;
 @property (nonatomic,copy) NSString * PY_PDTYEARTYPE;
 @property (nonatomic,copy) NSString * PY_PDTYEARNA;
 @property (nonatomic,copy) NSString * PY_RATECODE;
 @property (nonatomic,copy) NSString * PY_MINAMT;
 @property (nonatomic,copy) NSString * PY_MAXAMT;
-@property (nonatomic,assign) float PY_AMTUNIT;
-@property (nonatomic,assign) float PY_MINAGE;
-@property (nonatomic,assign) float PY_MAXAGE;
-@property (nonatomic,assign) float PY_INSUREYEAR;
+@property (nonatomic,assign) double PY_AMTUNIT;
+@property (nonatomic,assign) double PY_MINAGE;
+@property (nonatomic,assign) double PY_MAXAGE;
+@property (nonatomic,assign) double PY_INSUREYEAR;
 @property (nonatomic,copy) NSString * PY_INSUREYEARTYPE;
 @property (nonatomic,copy) NSString * PY_INSUREYEARNA;
 @property (nonatomic,copy) NSString * PY_WPCODE;
 @property (nonatomic,copy) NSString * PY_WP;
 @property (nonatomic,copy) NSString * PY_MINPREM;
-@property (nonatomic,assign) float PY_FYCITEM;
+@property (nonatomic,assign) double PY_FYCITEM;
 @property (nonatomic,copy) NSString * PY_MINAGE_IND;
 @property (nonatomic,copy) NSString * PY_MAXAGE_IND;
 @property (nonatomic,copy) NSString * MODIDATE;
